Add mingw-w64-ardour
First things first, so this pkgbuild is based from arch and biswa96 patches, I managed to build it. And it runs, but I have trouble with the license path. 
but hey it works?
![image](https://github.com/kreijstal-contributions/MINGW-packages/assets/2415206/5ac0a009-cba9-4f5c-a5cf-6a6b32fbbe81)

Cannot build on clang because there is no jack2 dependency in clang afaik. 
There is a bug in waf when it detects CC or CXX as env variables it attempts to get it as a file, it does not check the path for these, it treats it as absolute values.
I am not a musician, so no idea if it is functional.